moved Counter state to Counter folder
